Fix cluster check increase
When the cluster count in increased the charms workload status reports
that there are insufficent units which blocks the juju_set function.

This change adds the WAIT option to juju_set.py so the manifest can
stop the wait on workload status. The manifests have been updated to
set the WAIT option