Zaza Transition Phase 1
In phase one of the zaza transition we move all of the mojo_os_utils
and mojo_utils calls to zaza.utilities for those functions which
already exist in zaza. This leaves behind the functions which still
need to be migrated. But provides a bridge to begin using zaza in our
openstack mojo specs.

network_setup.py, bgp_speacker_setup.py and test_bgp_routes.py show the
way forward putting the brunt of the work on zaza and leaving only
imports on the mojo side.