Add boot from volume test
Starting OpenStack Ocata, the default mode of operation of
OpenStack Dashboard when launching an instance is to launch
it from a bootable volume. At the same version a new ceph-access
relation between cinder-ceph and nova-compute is required for
this to actually work.

Exercise this in the Mojo Specs to catch any regressions.