Stop testing rebooted gateway on xenial
In the vrrp tests limit the testing to checking connectivity
to the guest once the gateway housing the active router has been
shutdown. This omits testing that the shutdown gateway recovers
properly once it is brought back up. The reason for limiting the
tests is that when a xenial guest is rebooted the network
interface names may change which breaks the bridges. This issue
is specific to testing on top of OpenStack and should not occur
when using the maas provider.