Revert to testing ocata and earlier without vault
Revert to testing ocata and earlier without vault. This puts the
specs back into a known good position. SSL support can then be
added back in using the pre-generated certificate method which
is known to work with these releases or vault can be reintroduced
but that will involve fixing issues like Bug #1839019