Add HA and SSL support prior to Pike
This adds HA and SSL support to Mitaka, Newton and Ocata.

A limitation in Newton and earlier means that the IP SANS part
of cerificates is ignored. The impact of this is that neither
vault nor pre-deploy certificate generation will work. The Vault
generated certificates use IP SANS for the vip and the pre-deploy
generated certificates work by creating IP SANS entries for all
IPs in a network range of which the VIP is one. To work around this
Mitaka and Newton have there certificates generated and applied
straight after deployment. This is not ideal but I think it is a
reasonable compramise.

Additional Fixes for Mitaka -> Ocata:

* Add missing gnocchi <-> vault relation
* New bundle for use with Mitaka and Newton that does a full HA
  deploy without SSL as SSL is added afterwards
* Ensure designate workaround for vault is applied to Ocata
* expand_and_shrink_bindpy, validate_aodh.py and
  validate_designate.pya all needed fixes to pickup the local CA.